microarray: refactor 'annotate_probesets.py' to move into bcftbx.cli.
