+++ bin/annotate_probesets.py
@@ -1,0 +1,8 @@
+#!/usr/bin/env python
+#
+#     annotate_probesets.py: add descriptions to probe set names
+#     Copyright (C) University of Manchester 2021 Peter Briggs
+#
+from bcftbx.cli.annotate_probesets import main
+if __name__ == "__main__":
+     main()
